src/pipelines/option_trades/requirements.txt
attrs==24.2.0 ; python_version >= "3.11" and python_version < "3.14"
avro==1.12.0 ; python_version >= "3.11" and python_version < "3.14"
cattrs==24.1.2 ; python_version >= "3.11" and python_version < "3.14"
certifi==2024.8.30 ; python_version >= "3.11" and python_version < "3.14"
charset-normalizer==3.4.0 ; python_version >= "3.11" and python_version < "3.14"
click==8.1.7 ; python_version >= "3.11" and python_version < "3.14"
colorama==0.4.6 ; python_version >= "3.11" and python_version < "3.14" and platform_system == "Windows"
confluent-kafka[avro]==2.6.1 ; python_version >= "3.11" and python_version < "3.14"
dotenv-derive==0.1.0 ; python_version >= "3.11" and python_version < "3.14"
exchange-calendars==4.6 ; python_version >= "3.11" and python_version < "3.14"
fastavro==1.9.7 ; python_version >= "3.11" and python_version < "3.14"
idna==3.10 ; python_version >= "3.11" and python_version < "3.14"
korean-lunar-calendar==0.3.1 ; python_version >= "3.11" and python_version < "3.14"
numpy==2.2.0 ; python_version >= "3.11" and python_version < "3.14"
pandas-market-calendars==4.4.2 ; python_version >= "3.11" and python_version < "3.14"
pandas==2.2.3 ; python_version >= "3.11" and python_version < "3.14"
pyluach==2.2.0 ; python_version >= "3.11" and python_version < "3.14"
python-dateutil==2.9.0.post0 ; python_version >= "3.11" and python_version < "3.14"
python-dotenv==1.0.1 ; python_version >= "3.11" and python_version < "3.14"
pytz==2024.2 ; python_version >= "3.11" and python_version < "3.14"
requests==2.32.3 ; python_version >= "3.11" and python_version < "3.14"
six==1.17.0 ; python_version >= "3.11" and python_version < "3.14"
toolz==1.0.0 ; python_version >= "3.11" and python_version < "3.14"
tornado==6.4.2 ; python_version >= "3.11" and python_version < "3.14"
tradesignals @ file:///Users/macanderson/Tradesignals ; python_version >= "3.11" and python_version < "3.14"
tzdata==2024.2 ; python_version >= "3.11" and python_version < "3.14"
urllib3==2.2.3 ; python_version >= "3.11" and python_version < "3.14"
uvloop==0.21.0 ; python_version >= "3.11" and python_version < "3.14"